+++ src/hosted-fields/internal/components/label.js
@@ -4,7 +4,7 @@ function LabelComponent(options) {
   this.element = document.createElement('label');
 
   this.element.setAttribute('for', options.name);
-  this.element.innerHTML = options.label;
+  this.element.innerText = options.label;
 }
 
 module.exports = {