+++ src/hosted-fields/shared/browser-detection.js
@@ -3,9 +3,14 @@
 var isAndroid = require('@braintree/browser-detection/is-android');
 var isChromeOS = require('@braintree/browser-detection/is-chrome-os');
 var isIos = require('@braintree/browser-detection/is-ios');
+var isChrome = require('@braintree/browser-detection/is-chrome');
 
 function hasSoftwareKeyboard() {
   return isAndroid() || isChromeOS() || isIos();
+}
+
+function isChromeIos() {
+  return isChrome() && isIos();
 }
 
 module.exports = {
@@ -15,6 +20,7 @@ module.exports = {
   isIe10: require('@braintree/browser-detection/is-ie10'),
   isAndroid: isAndroid,
   isChromeOS: isChromeOS,
+  isChromeIos: isChromeIos,
   isFirefox: require('@braintree/browser-detection/is-firefox'),
   isIos: isIos,
   isIosWebview: require('@braintree/browser-detection/is-ios-webview'),