+++ src/lib/find-root-node.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = function findRootNode(element) {
+  while (element.parentNode) {
+    element = element.parentNode;
+  }
+
+  return element;
+};