+++ src/unionpay/shared/unionpay.js
@@ -11,7 +11,7 @@ var events = constants.events;
 var iFramer = require('@braintree/iframer');
 var methods = require('../../lib/methods');
 var VERSION = process.env.npm_package_version;
-var uuid = require('../../lib/vendor/uuid');
+var uuid = require('@braintree/uuid');
 var Promise = require('../../lib/promise');
 var wrapPromise = require('@braintree/wrap-promise');